refactor: convert project to ESM and use SVG icons sprite
This change required extracting some of the loaders into their own
modules so we could import them in `prime-content-cache.server.ts`
without running into issues with importing .svg files. For some reason,
`tsx` (which we need to run the server since it's in TS and imports TS)
was fine importing some .svg files but not all (even the same .svg file
imported in a different spot would cause problems).