feat: remove origin compression entirely
Fly.io's proxy automatically compresses the response unless you compress
it yourself. You can disable compression by setting the
`Content-Encoding` header to `identity`, but that affects CDNs that may
be in front of Fly's proxy, too. So, at this time, it seems the best
option is to remove compression from the origin server entirely, let
Fly's proxy compress the responses, and disable the CDN compression to
avoid double compression.